sixth commit and hopefully it's the last commit

+++ src/main/java/com/example/controller/StudentController.java
@@ -5,9 +5,8 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.ui.ModelMap;
+import org.springframework.web.bind.annotation.*;
 
 import com.example.model.StudentModel;
 import com.example.service.StudentService;
@@ -91,9 +90,63 @@ public class StudentController
     @RequestMapping("/student/delete/{npm}")
     public String delete (Model model, @PathVariable(value = "npm") String npm)
     {
-        studentDAO.deleteStudent (npm);
+        StudentModel student = studentDAO.selectStudent(npm);
+        if (student != null){
+            studentDAO.deleteStudent(npm);
+            return "delete";
+        }
+        else{
+            model.addAttribute("npm", npm);
+            return "not-found";
 
-        return "delete";
+        }
     }
 
+    @RequestMapping("/student/update/{npm}")
+    public String update (Model model,
+                          @PathVariable(value = "npm") String npm)
+    {
+
+        StudentModel student = studentDAO.selectStudent (npm);
+        if (student != null) {
+            model.addAttribute ("student", student);
+            return "form-update";
+        } else {
+            model.addAttribute ("npm", npm);
+            return "not-found";
+        }
+    }
+    @RequestMapping(value = "/student/update/submit", method = RequestMethod.POST)
+    public String updateSubmit (@ModelAttribute("student") StudentModel student, ModelMap model)
+    {
+        if(student.getName() == null || student.getName() == "") {
+            return "error";
+        }
+
+        StudentModel studentValid = studentDAO.selectStudent(student.getNpm());
+        if (studentValid != null) {
+            studentDAO.updateStudent(student.getNpm(), student.getName(), student.getGpa());
+            return "success-update";
+        }else {
+            model.addAttribute ("npm", student.getNpm());
+            return "not-found";
+        }
+    }
+
+    @RequestMapping(value = "/student/update/submit", method = RequestMethod.GET)
+    public String updateSubmitGET (@ModelAttribute("student") StudentModel student, ModelMap model)
+    {
+        if(student.getName() == null || student.getName() == "") {
+            return "error";
+        }
+
+        StudentModel studentValid = studentDAO.selectStudent(student.getNpm());
+        if (studentValid != null) {
+            studentDAO.updateStudent(student.getNpm(), student.getName(), student.getGpa());
+            return "success-update";
+        }else {
+            model.addAttribute ("npm", student.getNpm());
+            return "not-found";
+        }
+    }
 }
